Refactor build of performer to manage depenpencies more properly
