+++ Performers/project/Build.scala
@@ -18,86 +18,17 @@
 import sbt._
 import sbt.Keys._
 
-object BuildSettings {
+object ModuleDependencies {
 
-  val ParentProject = "jars_parent"
-  val Fey = "fey"
-  val Stream = "fey_stream"
-  val ZMQ = "fey_zmq"
+  import Dependencies._
 
-
-  val Version = "1.0"
-  val ScalaVersion = "2.11.8"
-
-  lazy val rootbuildSettings = Defaults.coreDefaultSettings ++ Seq(
-    name := ParentProject,
-    version := Version,
-    scalaVersion := ScalaVersion,
-    organization := "com.libit",
-    description := "Fey External Jars Project",
-    scalacOptions := Seq("-deprecation", "-unchecked", "-encoding", "utf8", "-Xlint")
-  )
-
-  lazy val FeybuildSettings = Defaults.coreDefaultSettings ++ Seq(
-    name := Fey,
-    version := Version,
-    scalaVersion := ScalaVersion,
-    organization := "com.litbit",
-    description := "Fey Development Instance",
-    scalacOptions := Seq("-deprecation", "-unchecked", "-encoding", "utf8", "-Xlint")
-  )
-
-  lazy val StreambuildSettings = Defaults.coreDefaultSettings ++ Seq(
-    name := Stream,
-    version := Version,
-    scalaVersion := ScalaVersion,
-    organization := "org.apache.iota.fey.performer",
-    description := "Simple Stream Application",
-    scalacOptions := Seq("-deprecation", "-unchecked", "-encoding", "utf8", "-Xlint")
-  )
-
-   lazy val ZMQbuildSettings = Defaults.coreDefaultSettings ++ Seq(
-    name := ZMQ,
-    version := Version,
-    scalaVersion := ScalaVersion,
-    organization := "org.apache.iota.fey.performer",
-    description := "ZMQ Application",
-    scalacOptions := Seq("-deprecation", "-unchecked", "-encoding", "utf8", "-Xlint")
-  )
-
+  val FeyDependencies = provided(akka_actor, fey)
+  val StreamDependencies = provided(akka_actor, fey)
+  val ZMQDependecies = provided(akka_actor, zmq, fey)
 }
 
-object Resolvers {
+object PerformersBuild extends Build {
 
-  val typesafe = "Typesafe Repository" at "http://repo.typesafe.com/typesafe/releases/"
-  val sonatype = "Sonatype Release" at "https://oss.sonatype.org/content/repositories/releases"
-  val mvnrepository = "MVN Repo" at "http://mvnrepository.com/artifact"
-  val litbitBitbucket = "Litbit Repo" at "https://s3-us-west-2.amazonaws.com/maven.litbit.com/snapshots"
-  val emuller = "emueller-bintray" at "http://dl.bintray.com/emueller/maven"
-
-  val allResolvers = Seq(typesafe, sonatype, mvnrepository, emuller, litbitBitbucket)
-
-}
-
-object Dependency {
-  //Use when building Artifacts (jar files)
-  val akka_actor = "com.typesafe.akka" %% "akka-actor" % "2.4.2" % "provided"
-  val fey = "org.apache.iota" %% "fey-core" % "1.0-SNAPSHOT" % "provided"
-  val zmq = "org.zeromq" % "jeromq" % "0.3.5"
-}
-
-object Dependencies {
-
-  import Dependency._
-
-  val FeyDependencies = Seq(akka_actor, fey)
-  val StreamDependencies = Seq(akka_actor, fey)
-  val ZMQDependecies = Seq(akka_actor, zmq, fey)
-}
-
-object JarsBuild extends Build {
-
-  import Resolvers._
   import BuildSettings._
 
   lazy val parent = Project(
@@ -112,50 +43,26 @@ object JarsBuild extends Build {
   lazy val fey = Project(
     id = "fey",
     base = file("./fey"),
-    settings = FeybuildSettings ++ Seq(
-      maxErrors := 5,
-      ivyScala := ivyScala.value map {
-        _.copy(overrideScalaVersion = true)
-      },
-      triggeredMessage := Watched.clearWhenTriggered,
-      resolvers := allResolvers,
-      libraryDependencies ++= Dependencies.FeyDependencies,
-      mainClass := Some("org.apache.iota.fey.Application"),
-      fork := true,
-      connectInput in run := true
+    settings = BasicSettings ++ FeybuildSettings ++ Seq(
+      libraryDependencies ++= ModuleDependencies.FeyDependencies,
+      mainClass := Some("org.apache.iota.fey.Application")
     ))
 
   lazy val stream = Project(
     id = "fey_stream",
     base = file("./stream"),
-    settings = StreambuildSettings ++ Seq(
-      maxErrors := 5,
+    settings = BasicSettings ++ StreambuildSettings ++ Seq(
+      libraryDependencies ++= ModuleDependencies.StreamDependencies,
+      mainClass := Some("org.apache.iota.fey.performer.stream.Application")
 
-      ivyScala := ivyScala.value map {
-        _.copy(overrideScalaVersion = true)
-      },
-      triggeredMessage := Watched.clearWhenTriggered,
-      resolvers := allResolvers,
-      libraryDependencies ++= Dependencies.StreamDependencies,
-      mainClass := Some("org.apache.iota.fey.performer.stream.Application"),
-      fork := true,
-      connectInput in run := true
     ))
 
    lazy val zmq = Project(
     id = "fey_zmq",
     base = file("./zmq"),
-    settings = ZMQbuildSettings ++ Seq(
-      maxErrors := 5,
-      ivyScala := ivyScala.value map {
-        _.copy(overrideScalaVersion = true)
-      },
-      triggeredMessage := Watched.clearWhenTriggered,
-      resolvers := allResolvers,
-      libraryDependencies ++= Dependencies.ZMQDependecies,
-      mainClass := Some("org.apache.iota.fey.performer.zmq.Application"),
-      fork := true,
-      connectInput in run := true
+    settings = BasicSettings ++ ZMQbuildSettings ++ Seq(
+      libraryDependencies ++= ModuleDependencies.ZMQDependecies,
+      mainClass := Some("org.apache.iota.fey.performer.zmq.Application")
     ))
 
 }
